restructure the header files, into a headers directory
